feat: Add custom 404 page
Add a new HTML file, 404.html, to the public directory. This file contains a custom 404 page with a message explaining that the specified file was not found on the website. Users are instructed to check the URL for mistakes and try again. This page was generated by the Firebase Command-Line Interface and can be modified by editing the 404.html file in the project's configured public directory.